# Conflicts:
#	logs/error.log
#	logs/log.2019-05-08.log
#	logs/log.log
#	src/main/resources/mapper/MeterMapper.xml
#	src/main/resources/mapper/UserMapper.xml

+++ src/main/java/com/iot/smart/water/meter/model/DailyData.java
@@ -2,7 +2,7 @@ package com.iot.smart.water.meter.model;
 
 public class DailyData {
     private String hour;
-    private float totalMilliters;
+    private long totalMilliters;
 
     public String getHour() {
         return hour;
@@ -12,11 +12,11 @@ public class DailyData {
         this.hour = hour;
     }
 
-    public float getTotalMilliters() {
+    public long getTotalMilliters() {
         return totalMilliters;
     }
 
-    public void setTotalMilliters(float totalMilliters) {
+    public void setTotalMilliters(long totalMilliters) {
         this.totalMilliters = totalMilliters;
     }
 }
